feat: add area label postion to show label in StackedArea
