+++ __tests__/integration/charts/population-by-state-area-normalize-stacked.ts
@@ -136,6 +136,14 @@ export function populationByStateAreaNormalizeStacked(): G2Spec {
           color: 'region',
           series: 'state',
         },
+        labels: [
+          {
+            text: 'state',
+            position: 'area',
+            selector: 'first',
+            fontSize: 10,
+          },
+        ],
       },
       {
         type: 'line',
